+++ src/App.jsx
@@ -10,11 +10,10 @@ import WorkArrangement from './Page/WorkArrangement/components/WorkArrangement';
 import ViewMyWork from './Page/WorkArrangement/components/ViewMyWork';
 import ViewMyWorkResult from './Page/WorkArrangement/components/ViewMyWorkResult';
 import Login from './Page/Login/components/Login';
-import SignUp from './Page/Login/SignUp/components/SignUp';
+import SignUp from './Page/Login/components/SignUp';
+import FindUser from './Page/Login/components/FindUser';
 import MyPage from './Page/MyPage/components/MyPage';
 import Notice from './Page/Notice/components/Notice';
-import FindID from './Page/Login/FindID/components/FindID';
-import FindPW from './Page/Login/FindPW/components/FindPW';
 import ModalCheck from './Global/components/Modal.components';
 import NoticeExample from './Page/Notice/components/NoticeExample';
 
@@ -22,7 +21,6 @@ import DevTerminal from './DevTerminal';
 import DevMockingApi from './Test/Mocking/API/DevMockingApi';
 import DevNotation from './Test/Mocking/API/DevNotation';
 import DevFetchTerminalTest from './DevFetchTerminalTest';
-import LoginTest from './Page/Login/components/Login.test';
 
 const App = () => {
   return (
@@ -43,17 +41,15 @@ const App = () => {
         <Route path="/ViewMyWorkResult" element={<ViewMyWorkResult />} />
         <Route path="/Login" element={<Login />} />
         <Route path="/SignUp" element={<SignUp />} />
+        <Route path="/FindUser" element={<FindUser />} />
         <Route path="/MyPage" element={<MyPage />} />
         <Route path="/Notice" element={<Notice />} />
-        <Route path="/FindID" element={<FindID />} />
-        <Route path="/FindPW" element={<FindPW />} />
         <Route path="/DevTerminal" element={<DevTerminal />} />
         <Route path="/DevMockingApi" element={<DevMockingApi />} />
         <Route path="/DevNotation" element={<DevNotation />} />
         <Route path="/NoticeExample" element={<NoticeExample />} />
         <Route path="/ModalCheck" element={<ModalCheck />} />
         <Route path="/DevFetch" element={<DevFetchTerminalTest />} />
-        <Route path="/LoginTest" element={<LoginTest />} />
       </Routes>
     </Router>
   );
