+++ src/DevTerminal.jsx
@@ -82,16 +82,13 @@ const Terminal = () => {
             <Link to="/MyPage">My Page</Link>
           </li>
           <li>
+            <Link to="/FindUser">FindUser</Link>
+          </li>
+          <li>
             <Link to="/Notation">Notation</Link>
           </li>
           <li>
             <Link to="/NotationExample">NotationExample</Link>
-          </li>
-          <li>
-            <Link to="/FindID">FindID</Link>
-          </li>
-          <li>
-            <Link to="/FindPW">FindPW</Link>
           </li>
           <li>
             <Link
