+++ src/Page/Login/components/SignUp.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import SignUpComponent from './SignUp.component';
+import '../css/SignUp.scss';
+
+const SignUp = () => {
+  return (
+    <div className="SignUp">
+      <SignUpComponent />
+    </div>
+  );
+};
+
+export default SignUp;
