+++ src/Page/Login/css/Findpw.component.scss
@@ -1,0 +1,106 @@
+.Findpw-container {
+  // display: flex;
+  // flex-direction: column;
+  // justify-content: center;
+  // align-items: center;
+  // min-width: 650px;
+  /* 최소 높이를 뷰포트 높이에 맞추기 */
+  // min-height: 700px;
+  // padding: 15px 70px;
+
+  .Findpw-group {
+    font-size: 16px;
+    font-weight: 400;
+    line-height: 24px;
+    color: #767676;
+    margin-top: 20px;
+
+    .Findpw-id,
+    .Findpw-password,
+    .Findpw-email,
+    .Findpw-name,
+    .Findpw-nickname {
+      display: flex;
+      flex-direction: column;
+
+      #Findpw-input-id,
+      #Findpw-input-password,
+      #Findpw-input-email,
+      #Findpw-input-name,
+      #Findpw-input-nickname {
+        width: calc(500px - 16px - 17.6px);
+        padding: 16px;
+        border: 1px solid #e5e5ec;
+        margin-top: 5px;
+      }
+    }
+
+    .Findpw-password,
+    .Findpw-email,
+    .Findpw-name,
+    .Findpw-nickname {
+      span {
+        margin-top: 10px;
+      }
+    }
+  }
+
+  .Findpw-submit {
+    #Findpw-button {
+      border: none;
+      width: 500px;
+      background-color: #204598;
+      color: #ffffff;
+      margin-bottom: 24px;
+      margin-top: 40px;
+      padding: 16px;
+    }
+
+    .Findpw-link {
+      display: flex;
+      justify-content: center;
+
+      #Findpw-find {
+        color: #767676;
+        margin-right: 16px;
+      }
+
+      #Findpw-Findpw {
+        font-size: 16px;
+        font-weight: 400;
+        line-height: 24px;
+      }
+
+    }
+  }
+}
+
+/* 뷰포트 높이가 900px 이상일 때 `min-height`를 1000px로 변경 */
+@media (min-height: 1000px) {
+  .Findpw-container {
+    min-width: 500px;
+    // min-height: calc(1000px - 120px);
+    // padding: 60px 150px;
+  }
+
+  .Findpw-container .Findpw-submit {
+    // margin-top: calc(140px - 84.79px);
+    margin-top: 140px;
+  }
+
+  .Findpw-container .Findpw-group {
+    margin-top: 60px;
+  }
+
+  .Findpw-container .Findpw-group .Findpw-id #Findpw-input-id {
+    margin-top: 12px;
+  }
+
+  .Findpw-container .Findpw-group .Findpw-password span {
+    margin-top: 24px;
+  }
+
+  .Findpw-container .Findpw-group .Findpw-password #Findpw-input-password {
+    margin-top: 12px;
+  }
+}
