+++ src/Page/Login/css/SignUp.component.scss
@@ -1,0 +1,117 @@
+.SignUp-container {
+  background-color: white;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  // min-width: 650px;
+  /* 최소 높이를 뷰포트 높이에 맞추기 */
+  min-height: 700px;
+  padding: 50px 70px;
+
+  #SignUp-Logo {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+
+    #SignUp-mention {
+      margin: 20px 0px 13px 0px;
+    }
+  }
+
+  .SignUp-group {
+    font-size: 16px;
+    font-weight: 400;
+    line-height: 24px;
+    color: #767676;
+    margin-top: 20px;
+
+    .SignUp-id,
+    .SignUp-password,
+    .SignUp-email,
+    .SignUp-name,
+    .SignUp-nickname {
+      display: flex;
+      flex-direction: column;
+
+      #SignUp-input-id,
+      #SignUp-input-password,
+      #SignUp-input-email,
+      #SignUp-input-name,
+      #SignUp-input-nickname {
+        width: calc(500px - 16px - 17.6px);
+        padding: 16px;
+        border: 1px solid #e5e5ec;
+        margin-top: 5px;
+      }
+    }
+
+    .SignUp-password,
+    .SignUp-email,
+    .SignUp-name,
+    .SignUp-nickname {
+      span {
+        margin-top: 10px;
+      }
+    }
+  }
+
+  .SignUp-submit {
+    #SignUp-button {
+      border: none;
+      width: 500px;
+      background-color: #204598;
+      color: #ffffff;
+      margin-bottom: 24px;
+      margin-top: 100px;
+      padding: 16px;
+    }
+
+    .SignUp-link {
+      display: flex;
+      justify-content: center;
+
+      #SignUp-find {
+        color: #767676;
+        margin-right: 16px;
+      }
+
+      #SignUp-SignUp {
+        font-size: 16px;
+        font-weight: 400;
+        line-height: 24px;
+      }
+
+    }
+  }
+}
+
+/* 뷰포트 높이가 900px 이상일 때 `min-height`를 1000px로 변경 */
+@media (min-height: 1000px) {
+  .SignUp-container {
+    min-width: 500px;
+    min-height: calc(1000px - 120px);
+    padding: 60px 150px;
+  }
+
+  .SignUp-container .SignUp-submit {
+    margin-top: 140px;
+  }
+
+  .SignUp-container .SignUp-group {
+    margin-top: 60px;
+  }
+
+  .SignUp-container .SignUp-group .SignUp-id #SignUp-input-id {
+    margin-top: 12px;
+  }
+
+  .SignUp-container .SignUp-group .SignUp-password span {
+    margin-top: 24px;
+  }
+
+  .SignUp-container .SignUp-group .SignUp-password #SignUp-input-password {
+    margin-top: 12px;
+  }
+}
