+++ src/Page/Login/css/SignUp.scss
@@ -1,4 +1,4 @@
-.Login {
+.SignUp {
   // background-color: aqua;
   display: flex;
   justify-content: center;
